move meet actions to meet.ts, deleted tournaments in prisma schema, increase number of possible participants in meet, change revalidatePath
