+++ actions/meet.ts
@@ -2,6 +2,7 @@
 import { prisma } from "@/lib/db";
 import { revalidatePath } from "next/cache";
 import { isFuture, isToday, format } from "date-fns";
+import { Tag } from "@prisma/client";
 
 // Helper function to check if a given time is in the future
 function isTimeInFuture(time: string) {
@@ -12,6 +13,57 @@ function isTimeInFuture(time: string) {
   // Compare meet time with current time
   return meetTimeNumber > timeNowNumber;
 }
+
+// creating a meet / session
+interface MeetProps {
+  date: Date;
+  time: string;
+  duration: number;
+  isPublic: boolean;
+  creatorId: string;
+  guests: number;
+  notes?: string;
+  venueId: string;
+  activityTypeName: string;
+}
+
+export const createMeet = async ({
+  date,
+  time,
+  duration,
+  isPublic,
+  creatorId,
+  guests,
+  notes,
+  venueId,
+  activityTypeName,
+}: MeetProps) => {
+  await prisma.meet.create({
+    data: {
+      date: date,
+      time: time,
+      duration: duration,
+      isPublic: isPublic,
+      creator: {
+        connect: {
+          id: creatorId,
+        },
+      },
+      Venue: {
+        connect: {
+          id: venueId,
+        },
+      },
+      activityType: {
+        connect: {
+          name: activityTypeName,
+        },
+      },
+      guests: guests,
+      notes: notes,
+    },
+  });
+};
 
 // Main function to delete a meet
 export async function deleteMeet(meetId: string, userId: string) {
@@ -50,7 +102,7 @@ export async function deleteMeet(meetId: string, userId: string) {
     });
 
     // Revalidate the path to update the UI
-    revalidatePath("/meets");
+    revalidatePath("/meet");
 
     // Return success message
     return { success: true, message: "Meet deleted successfully" };
@@ -64,3 +116,35 @@ export async function deleteMeet(meetId: string, userId: string) {
     };
   }
 }
+
+// Tag creation on meet
+
+type Props = {
+  names: string[];
+  tag?: Tag;
+};
+// update new tags that are added by the user by checking if it exists and if not add it to db
+
+export const updateTags = async ({ names, tag }: Props) => {
+  const exisitngTags = await prisma.tag.findMany({ select: { name: true } });
+
+  const exisitngTagNames = new Set(exisitngTags.map((t) => t.name));
+
+  const newTags = names.filter((name) => !exisitngTagNames.has(name));
+
+  if (newTags.length > 0) {
+    await prisma.tag.createMany({
+      data: newTags.map((name) => ({ name })),
+      skipDuplicates: true,
+    });
+  }
+
+  if (tag && tag.name) {
+    await prisma.tag.upsert({
+      where: { name: tag.name },
+      update: tag,
+      create: tag,
+    });
+  }
+  console.log("Tag ${tag.name} updatetd or created successfully");
+};
